% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{generateQCreport_table}
\alias{generateQCreport_table}
\title{genereateQCreport_table}
\usage{
generateQCreport_table(
  seu_obj = NULL,
  features = NULL,
  pdfFile = "QCReportTable.pdf",
  sample_id = NULL,
  path = NULL,
  platform = NULL
)
}
\arguments{
\item{seu_obj}{A Seurat object, which must have properties like \code{path}, \code{platform}, and specific assay data.
Must inherit from "Seurat".}

\item{features}{Optional; A character vector of features to include in the analysis.
Defaults to all features in \code{seu_obj} if \code{NULL}.}

\item{pdfFile}{A character string specifying the path and name of the output PDF file.
Defaults to "QCReportTable.pdf".}

\item{sample_id}{Identifier for the sample being processed.}

\item{path}{The file path from which to load the spatial transcriptomics data.}

\item{platform}{The platform from which the data originates. Valid options are 'Xenium', 'CosMx',
and 'Merscope'. Note: 'Merscope' is currently not supported.}
}
\value{
A data frame with the QC metrics for the provided Seurat object, which is also printed to the console.
Additionally, a PDF file is generated containing a plot and table of the metrics.
}
\description{
Generate QC Report Table for Seurat Object
}
\details{
Generate a QC report for a given Seurat object, allowing users to specify features and
output to a PDF file. It provides various statistical analyses including entropy, sparsity,
transcript counts per area, transcript counts per cell, and signal ratios. It also adjusts its behavior
based on the specified platform within the Seurat object (Xenium, CosMx, or Merscope).

Calculate various metrics such as entropy and sparsity using the \code{BioQC} and \code{coop} packages.
It adjusts its processing logic based on the 'platform' attribute of the Seurat object to handle data from
different technologies like Xenium, CosMx, or Merscope. It stops with an error if the object is not a Seurat object.
The results are plotted using \code{ggplot2} and \code{gridExtra} for layout adjustments.
}
