% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{getClusterMetrics}
\alias{getClusterMetrics}
\title{getClusterMetrics}
\usage{
getClusterMetrics(seu_obj, metadata_col)
}
\arguments{
\item{seu_obj}{A Seurat object for which clustering metrics are to be calculated.}

\item{metadata_col}{The metadata column in \code{seu_obj} that contains cell type predictions for comparison against clustering results.}
}
\value{
A data frame with sample identifiers, platforms, and the calculated ARI and NMI metrics.
}
\description{
Applies the clustering analysis to a Seurat object (seu_obj) to calculate the Adjusted Rand Index (ARI) and Normalized Mutual Information (NMI)
between the resulting clusters and a specified cell type prediction.
ARI and NMI are metrics used to assess the quality of clustering and the agreement between predicted cell types and clustering outcomes.
}
\details{
Clustering is performed on the \code{seu_obj} data, which is first normalized and scaled. PCA is then run, followed by neighbor finding
and cluster identification. The ARI and NMI are calculated to evaluate the clustering quality, with ARI measuring the similarity
between two data clusterings and NMI providing a normalized measure of the mutual dependence between the predicted cell types and
the clusters. These metrics offer insights into the coherence of cell type predictions and the overall quality of the clustering process.
}
