% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{getCorrelation}
\alias{getCorrelation}
\title{getCorrelation}
\usage{
getCorrelation(seu_obj, ref)
}
\arguments{
\item{seu_obj}{A Seurat object containing spatial transcriptomics data.}

\item{ref}{A Seurat object containing single-cell RNA-seq data used as a reference.}
}
\value{
A data frame with columns for \code{sample_id}, \code{platform}, and the Spearman correlation coefficient \code{value}.
}
\description{
Identifies common genes between a spatial transcriptomics dataset (seu_obj) and a single-cell RNA-seq reference dataset (ref).
From this objec  the Spearman correlation coefficient is computed based on the mean expression levels of these genes across the datasets.
The result is a single correlation coefficient value that quantifies the similarity between the two datasets at the expression level of shared genes.
}
\details{
This function is particularly useful for assessing the overall concordance between spatial transcriptomics data and a reference single-cell RNA-seq dataset.
By focusing on common genes, it provides a metric of similarity that can help validate spatial data against established single-cell profiles or compare spatial datasets against different references.
}
