% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{getGlobalFDR}
\alias{getGlobalFDR}
\title{getGlobalFDR}
\usage{
getGlobalFDR(
  seu_obj = NULL,
  features = NULL,
  tx_file = "path_to_txFile",
  cellSegMeta = "path_to_cellMeta",
  platform = NULL,
  path
)
}
\arguments{
\item{seu_obj}{A Seurat object containing spatial data, including a path and platform attribute.}

\item{features}{An optional vector of feature names (gene names) for which to calculate the FDR.
If NULL (default), FDR is calculated for all features in the seu_obj.}

\item{tx_file}{Path to transcription file.}

\item{platform}{The platform from which the data originates. Valid options are 'Xenium', 'CosMx',
and 'Merscope'. Note: 'Merscope' is currently not supported.}

\item{path}{Path to generate tx file.}
}
\value{
A data frame with columns for sample_id, platform, and the
calculated mean FDR across the specified features.
}
\description{
Calculate specificity as Global False Discovery Rate (FDR).
}
\details{
Computes the global FDR for specified features (or all features by default) in a Seurat object.
It leverages transcript localization data to calculate FDR based on the proportion of negative control
or blank barcode expressions compared to the expression of actual genes.
}
