% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{getMaxRatio}
\alias{getMaxRatio}
\title{getMaxRatio}
\usage{
getMaxRatio(
  seu_obj = NULL,
  features = NULL,
  expMat = "path_to_expMat",
  platform = NULL
)
}
\arguments{
\item{seu_obj}{A Seurat object including assays for target genes ('RNA')
and control probes ('ControlProbe').}

\item{features}{An optional vector of gene identifiers for which to perform the calculation. If NULL, the calculation
includes all genes in the RNA assay.}

\item{expMat}{Path to exprMatrix file.}

\item{platform}{The platform from which the data originates. Valid options are 'Xenium', 'CosMx',
and 'Merscope'. Note: 'Merscope' is currently not supported.}
}
\value{
A data frame with \code{sample_id}, \code{platform}, and the calculated \code{value} of the log-ratio.
}
\description{
Calculate the Log-ratio of highest mean expression vs. mean noise.
}
\details{
The function identifies the maximum mean expression value among the specified features
(or all features if none are specified) and calculates its log-ratio to the mean expression value
of negative control probes. This log-ratio reflects the dynamic range of the dataset, indicating
the spread between the highest signal and background noise levels, which is critical for assessing
data quality and sensitivity of detection in spatial transcriptomics.
}
