% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{getMeanExpression}
\alias{getMeanExpression}
\title{getMeanExpression}
\usage{
getMeanExpression(
  seu_obj = NULL,
  features = NULL,
  expMat = "path_to_expMatrix",
  platform = NULL
)
}
\arguments{
\item{seu_obj}{A Seurat object.
This object must have an RNA assay for target genes and a ControlProbe assay for control probes.}

\item{features}{An optional vector of gene names for which to calculate mean expression.
If NULL, mean expression is calculated for all genes in the RNA assay.}

\item{expMat}{Path to exprMatrix file.}

\item{platform}{The platform from which the data originates. Valid options are 'Xenium', 'CosMx',
and 'Merscope'. Note: 'Merscope' is currently not supported.}
}
\value{
A data frame with columns \code{target} (gene or control probe name), \code{value} (mean expression level),
\code{type} (indicating whether the row represents a Gene or Control), \code{platform}, and \code{sample_id}.
}
\description{
It calculated mean expression per probe.
}
\details{
Computes the mean expression levels for a specified list of features (genes)
and control probes within a Seurat object. This function separately calculates the mean
expression for both target genes and control probes, then combines these results into a
single data frame. If no features are specified, it defaults to all genes in the RNA assay.
}
