% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{getMeanSignalRatio}
\alias{getMeanSignalRatio}
\title{getMeanSignalRatio}
\usage{
getMeanSignalRatio(
  seu_obj = NULL,
  features = NULL,
  platform = NULL,
  expMat = "path_to_expMat"
)
}
\arguments{
\item{seu_obj}{A Seurat object, expected to have both 'RNA' and 'ControlProbe'
assays for calculating mean expression levels of genes and negative control probes, respectively.}

\item{features}{An optional vector of gene names for which to calculate the mean log-ratio. If NULL, the calculation
is performed for all genes present in the 'RNA' assay of the seu_obj.}
}
\value{
A data frame with three columns: \code{sample_id}, \code{platform}, and \code{value}, where \code{value} represents the calculated
mean log-ratio of gene expression to negative control probe expression for the selected genes. This summary can be
used to assess the signal-to-noise ratio in the dataset.
}
\description{
Calcuates the log-ratio of mean gene expression counts to mean negative probe counts.
}
\details{
Computes the log-ratio of the mean expression levels of specified genes (or all genes if none are specified)
to the mean expression levels of negative control probes within a Seurat object. This metric can provide insights into
the overall signal strength relative to background noise levels.
}
