% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{getMECR}
\alias{getMECR}
\title{getMECR}
\usage{
getMECR(seu_obj = NULL, expMat = "path_to_expMat", platform = NULL)
}
\arguments{
\item{seu_obj}{A seurat object.}

\item{expMat}{Path to exprMatrix file.}

\item{platform}{The platform from which the data originates. Valid options are 'Xenium', 'CosMx',
and 'Merscope'. Note: 'Merscope' is currently not supported.}
}
\value{
A data frame containing the \code{sample_id}, \code{platform}, and the computed
MECR value. The MECR value is rounded to three decimal places and represents
the average mutually exclusive co-expression rate across the selected markers.
}
\description{
Calculate Mutually Exclusive Co-expression Rate (MECR) which is a metric used for determining specificity.
}
\details{
Based on a predefined set of markers and their associated cell types, this function computes the rate
at which pairs of markers are expressed in mutually exclusive patterns within cells. The calculation is limited
to markers found in the dataset and can be adjusted to focus on a subset by specifying it. The result is a
single MECR value that quantifies the overall mutually exclusive co-expression pattern, rounded to three decimal places.
}
