% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{getSilhouetteWidth}
\alias{getSilhouetteWidth}
\title{getSilhouetteWidth}
\usage{
getSilhouetteWidth(seu_obj)
}
\arguments{
\item{seu_obj}{A Seurat object.}
}
\value{
A data frame containing the \code{sample_id}, \code{platform},
and the average silhouette width across all clusters, rounded to three decimal places.
}
\description{
Calculate the silhouette width which provides a metric for evaluating consistency within data clusters.
}
\details{
The silhouette width calculation involves preprocessing steps including normalization and scaling of the data, PCA for dimensionality reduction, and clustering.
The silhouette width is then calculated for a downsampled subset of cells to ensure computational efficiency.
This metric helps in assessing the cohesion and separation of the identified clusters, with higher values indicating better defined clusters.
}
