% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{getSparsity}
\alias{getSparsity}
\title{getSparsity}
\usage{
getSparsity(
  seu_obj = NULL,
  features = NULL,
  expMat = "path_to_expMat",
  platform = NULL
)
}
\arguments{
\item{seu_obj}{A Seurat object.}

\item{features}{An optional vector of feature names (e.g., gene symbols) to include in the calculation.
Defaults to NULL, in which case the calculation uses all available features in the RNA assay
of the Seurat object.}

\item{expMat}{Path to exprMatrix file.}

\item{platform}{The platform from which the data originates. Valid options are 'Xenium', 'CosMx',
and 'Merscope'. Note: 'Merscope' is currently not supported.}
}
\value{
A data frame.
}
\description{
Calculate the sparsity (as a count or proportion) of a gene expression count matrix.
}
\details{
For example, .99 sparsity means 99\% of the values are zero. Similarly, a sparsity of 0 means the matrix is fully dense.
}
