% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotCellCount}
\alias{plotCellCount}
\title{plotCellCount}
\usage{
plotCellCount(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
Plot.
}
\description{
Computes and compares cell type proportions between spatial and reference datasets.
}
