% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotCellTypeProportion}
\alias{plotCellTypeProportion}
\title{plotCellTypeProportion}
\usage{
plotCellTypeProportion(df)
}
\arguments{
\item{df}{Data frame with \code{group}, \code{prop}, \code{cell_type}, and \code{sample_id} columns.}
}
\value{
A ggplot object visualizing the proportion of each cell type within and across samples.
}
\description{
Creates bar plots to show the proportion of each cell type within samples.
}
