% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{plotCorrelation}
\usage{
plotCorrelation(df)
}
\arguments{
\item{df}{Data frame with \code{sample_id}, \code{value_sample}, and \code{value_ref} columns indicating sample IDs and expression values in the spatial and reference datasets, respectively.}
}
\value{
A ggplot object showing correlation scatter plots for each sample, with log-transformed axes.
}
\description{
Creates scatter plots to visualize correlation between spatial and reference dataset expression values across samples.
}
