% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotFractionTxInCell}
\alias{plotFractionTxInCell}
\title{plotFractionTxInCell}
\usage{
plotFractionTxInCell(df)
}
\arguments{
\item{df}{Data frame with \code{sample_id} and \code{value} columns indicating the fraction of transcripts in cells.}
}
\value{
A ggplot object depicting the fraction of transcripts within cells across samples.
}
\description{
Generates a bar plot showing the fraction of transcripts located within cells for each sample.
}
