% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotMaxExpression}
\alias{plotMaxExpression}
\title{plotMaxExpression}
\usage{
plotMaxExpression(df)
}
\arguments{
\item{df}{Data frame with \code{sample_id}, \code{value}, and \code{platform} columns.}
}
\value{
A ggplot object illustrating maximal gene detection rates per cell, differentiated by platform.
}
\description{
Creates a jitter and boxplot showing the maximal gene detection rate per cell across samples and platforms.
}
