% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotMeanExpression}
\alias{plotMeanExpression}
\title{plotMeanExpression}
\usage{
plotMeanExpression(df)
}
\arguments{
\item{df}{Data frame with \code{sample_id}, \code{value}, and \code{type} columns.}
}
\value{
A ggplot object depicting mean gene detection rates per cell, differentiated by cell type.
}
\description{
Generates a jitter and boxplot visualizing the mean gene detection rate per cell across samples and cell types.
}
