% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotMetrics}
\alias{plotMetrics}
\title{plotMetrics}
\usage{
plotMetrics(
  metrics_list = NULL,
  PlotAutocorr = T,
  PlotSparsity = T,
  PlotEntropy = T,
  PlotClusterSilhouette = T,
  ncol = 14,
  nrow = 1,
  rel_widths = c(0.75, 0.4, 0.5, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.75, 0.75, 0.8, 0.3,
    0.8, 0.8)
)
}
\arguments{
\item{metrics_list}{A list containing the metrics results.}

\item{PlotAutocorr}{Logical; if \code{TRUE}, plots the autocorrelation metric. Default is \code{TRUE}.}

\item{PlotSparsity}{Logical; if \code{TRUE}, plots the sparsity metric;Default is \code{TRUE}.}

\item{PlotEntropy}{Logical; if \code{TRUE}, plots the entropy metric; Default is \code{TRUE}.}

\item{PlotClusterSilhouette}{Logical; if \code{TRUE}, plots the cluster silhouette metric; Default is \code{TRUE}.}

\item{ncol}{Integer; the number of columns in the plot grid; Default is 14.}

\item{nrow}{Integer; the number of rows in the plot grid; Default is 1.}

\item{rel_widths}{Numeric vector; relative widths of columns in the plot grid.}
}
\value{
A plot object created with cowplot::plot_grid.
}
\description{
Generates plots from technical metrics in spatial touchstone.
}
