% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotMorans}
\alias{plotMorans}
\title{plotMorans}
\usage{
plotMorans(df)
}
\arguments{
\item{df}{Data frame with \code{sample_id}, \code{value}, and \code{type} columns for Moran's I values and sample types.}
}
\value{
A ggplot object depicting Moran's I spatial autocorrelation values, differentiated by sample type.
}
\description{
Creates a plot for Moran's I spatial autocorrelation values across samples.
}
