% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotSampleLabel}
\alias{plotSampleLabel}
\title{plotSampleLabel}
\usage{
plotSampleLabel(sample_meta)
}
\arguments{
\item{sample_meta}{sample_meta}
}
\value{
Data frame with cell type-specific correlation values.
}
\description{
Performs Spearman correlation analysis for each cell type between spatial and reference datasets.
}
