% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotSignalRatio}
\alias{plotSignalRatio}
\title{plotSignalRatio}
\usage{
plotSignalRatio(df)
}
\arguments{
\item{df}{Data frame with \code{sample_id} and \code{value} columns indicating the mean log10-ratio of expression over noise.}
}
\value{
A ggplot object visualizing the signal-to-noise ratio for each sample.
}
\description{
Creates a bar plot for the signal-to-noise ratio of gene expression across samples.
}
