% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotTxPerArea}
\alias{plotTxPerArea}
\title{plotTxPerArea}
\usage{
plotTxPerArea(df)
}
\arguments{
\item{df}{Data frame with \code{sample_id} and \code{value} columns indicating sample IDs and the number of transcripts per unit area, respectively.}
}
\value{
A ggplot object depicting the distribution of transcripts per unit area for each sample.
}
\description{
Generates a scatter plot to show the number of transcripts per unit area across samples.
}
