% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotTxPerCell}
\alias{plotTxPerCell}
\title{plotTxPerCell}
\usage{
plotTxPerCell(df)
}
\arguments{
\item{df}{Data frame with \code{sample_id} and \code{value} columns indicating sample IDs and the number of transcripts per cell, respectively.}
}
\value{
A ggplot object visualizing the number of transcripts per cell for each sample.
}
\description{
Creates a scatter plot visualizing the number of transcripts per cell across samples.
}
