% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_final.R
\name{plotTxPerNuc}
\alias{plotTxPerNuc}
\title{Plot Transcripts Per Nucleus}
\usage{
plotTxPerNuc(df)
}
\arguments{
\item{df}{Data frame with \code{sample_id} and \code{value} columns for sample IDs and the number of transcripts per nucleus, respectively.}
}
\value{
A ggplot object illustrating the transcripts count per nucleus across samples.
}
\description{
Produces a scatter plot that displays the number of transcripts per nucleus for each sample.
}
